# backend/src/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "E-commerce API",
    "description": "Essa é uma API simples para o gerenciamento de banco de um sistema de gerenciamento de compras de clientes de um comércio. O sistema é responsável por criar e gerenciar clientes, produtos e compras.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "tags": [
    {
      "name": "client",
      "description": "Operações sobre os clientes"
    },
    {
      "name": "order",
      "description": "Operações sobre as compras"
    },
    {
      "name": "product",
      "description": "Operações sobre os produtos"
    }
  ],
  "paths": {
    "/client": {
      "get": {
        "tags": [
          "client"
        ],
        "summary": "Buscar todos clientes",
        "description": "Retorna todos clientes armazenados no banco",
        "operationId": "getClients",
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "client"
        ],
        "summary": "Criar cliente",
        "description": "Adiciona um novo cliente no comércio",
        "operationId": "createCliente",
        "requestBody": {
          "description": "Cria um novo cliente no banco",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cliente criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/client/{id}": {
      "put": {
        "tags": [
          "client"
        ],
        "summary": "Atualizar cliente",
        "description": "Atualiza o cliente com o id dado",
        "operationId": "updateClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do cliente",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Atualiza um cliente no banco",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "client"
        ],
        "summary": "Remover cliente",
        "description": "Remove o cliente com o id dado",
        "operationId": "removeClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do cliente",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Buscar todos produtos",
        "description": "Retorna todos produtos armazenados no banco",
        "operationId": "getProducts",
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Criar produto",
        "description": "Adiciona um novo produto no comércio",
        "operationId": "createProduct",
        "requestBody": {
          "description": "Cria um novo produto no banco",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Produto criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/product/{id}": {
      "put": {
        "tags": [
          "product"
        ],
        "summary": "Atualizar produto",
        "description": "Atualiza o produto com o id dado",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Atualiza um produto no banco",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Remover produto",
        "description": "Remove o produto com o id dado",
        "operationId": "removeProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/order": {
      "get": {
        "tags": [
          "order"
        ],
        "summary": "Buscar todas compras",
        "description": "Retorna todas compras armazenadas no banco",
        "operationId": "getOrders",
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "order"
        ],
        "summary": "Criar compra",
        "description": "Adiciona uma nova compra no comércio",
        "operationId": "createOrder",
        "requestBody": {
          "description": "Cria uma nova compra no banco",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Compra criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/order/{id}": {
      "put": {
        "tags": [
          "order"
        ],
        "summary": "Atualizar compra",
        "description": "Atualiza a compra com o id dado",
        "operationId": "updateOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da compra",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Cria uma nova compra no banco",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "order"
        ],
        "summary": "Remover compra",
        "description": "Remove a compra com o id dado",
        "operationId": "removeOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "cliente_id": {
            "type": "integer",
            "description": "Id do cliente que realizou a compra"
          },
          "produto_id": {
            "type": "integer",
            "description": "Id do produto que foi comprado"
          },
          "quantidade": {
            "type": "integer",
            "example": 3
          },
          "preco_total": {
            "type": "number",
            "description": "Preço total da compra",
            "example": 10
          },
          "data": {
            "type": "string",
            "format": "date",
            "description": "Data da compra"
          }
        },
        "xml": {
          "name": "order"
        }
      },
      "GetOrder": {
        "type": "object",
        "properties": {
          "cliente": {
            "type": "string",
            "description": "Nome do cliente que realizou a compra"
          },
          "produto": {
            "type": "string",
            "description": "Nome do produto que foi comprado"
          },
          "quantidade": {
            "type": "integer",
            "example": 3
          },
          "preco_total": {
            "type": "number",
            "description": "Preço total da compra",
            "example": 10
          },
          "data": {
            "type": "string",
            "format": "date",
            "description": "Data da compra"
          }
        },
        "xml": {
          "name": "getorder"
        }
      },
      "Client": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Maria Luiza"
          },
          "email": {
            "type": "string",
            "example": "mluiza@gmail.com"
          }
        },
        "xml": {
          "name": "cliente"
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Caneta"
          },
          "preco": {
            "type": "number",
            "example": 21.5
          }
        },
        "xml": {
          "name": "produto"
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Response": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "OrderResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Response"
          },
          {
            "properties": {
              "result": {
                "type": "array",
                "items": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Order"
                    },
                    {
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "example": 1
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        ]
      },
      "GetOrderResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Response"
          },
          {
            "properties": {
              "result": {
                "type": "array",
                "items": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GetOrder"
                    },
                    {
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "example": 1
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        ]
      },
      "ClientResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Response"
          },
          {
            "properties": {
              "result": {
                "type": "array",
                "items": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Client"
                    },
                    {
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "example": 1
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        ]
      },
      "ProductResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Response"
          },
          {
            "properties": {
              "result": {
                "type": "array",
                "items": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Product"
                    },
                    {
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "example": 1
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        ]
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}